+++ spec/lib/water_drop/producer_proxy_spec.rb
@@ -1,5 +1,3 @@
-require 'spec_helper'
-
 RSpec.describe WaterDrop::ProducerProxy do
   subject { described_class }