+++ spec/lib/water_drop/version_spec.rb
@@ -1,5 +1,3 @@
-require 'spec_helper'
-
 RSpec.describe WaterDrop do
   it { expect { WaterDrop::VERSION }.not_to raise_error }
 end