+++ spec/lib/water_drop_spec.rb
@@ -1,4 +1,3 @@
-require 'spec_helper'
 require 'logger'
 
 RSpec.describe WaterDrop do